restructure test directory
