+++ test/util/exportToNDJson.test.js
@@ -1,15 +1,15 @@
-const build = require('../src/server/app');
-const { exportToNDJson, patientsQueryForType, getDocuments } = require('../src/util/exportToNDJson');
-const { cleanUpDb, createTestResourceWithConnect } = require('./populateTestData');
-const testPatient = require('./fixtures/testPatient.json');
-const testEncounter = require('./fixtures/testEncounter.json');
+const build = require('../../src/server/app');
+const { exportToNDJson, patientsQueryForType, getDocuments } = require('../../src/util/exportToNDJson');
+const { cleanUpDb, createTestResourceWithConnect } = require('../populateTestData');
+const testPatient = require('../fixtures/testPatient.json');
+const testEncounter = require('../fixtures/testEncounter.json');
 
-const testServiceRequest = require('./fixtures/testServiceRequest.json');
+const testServiceRequest = require('../fixtures/testServiceRequest.json');
 const app = build();
 const fs = require('fs');
 // import queue to close open handles after tests pass
 // TODO: investigate why queues are leaving open handles in this file
-const queue = require('../src/resources/exportQueue');
+const queue = require('../../src/resources/exportQueue');
 const mockType = ['Patient'];
 
 const expectedFileName = './tmp/123456/Patient.ndjson';
